+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     - `inv::Ix`
     - `inv::InvInBlock` and `inv::IconRow` now have public `new()` functions.
     - `math::FaceMap::sum()`
+    - New universe member type `sound::SoundDef` allows defining synthesized sounds as game content.
     - `impl IntoIterator for math::FaceMap`
     - `math::PositiveSign::INFINITY`